% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aaindex1.R
\docType{data}
\name{aaindex1}
\alias{aaindex1}
\title{List of 571 Amino Acid Physicochemical Indexes from AAindex Database}
\format{
\code{\link{AutomorphismList}} class object.
}
\usage{
aaindex1
}
\description{
The aminoacid indexes from Amino Acid Index Database
\url{https://www.genome.jp/aaindex/} are provided here. AAindex (ver.9.2)
is a database of numerical indices representing various physicochemical and
biochemical properties of amino acids and pairs of amino acids.
}
\examples{
## Load the mutation matrices from database from the packages
data(aaindex1, package = "GenomAutomorphism")

## Get the available mutation matrices
mat <- aa_mutmat(aaindex = aaindex1, acc_list = TRUE)
mat[1:10]
}
\seealso{
\link{aaindex2} and \link{aaindex3}.
}
\author{
Robersy Sanchez \url{https://genomaths.com}
}
\keyword{datasets}
