% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{as.AutomorphismList}
\alias{as.AutomorphismList}
\alias{as.AutomorphismList,GRangesList,GRanges_OR_NULL-method}
\alias{as.AutomorphismList,list,GRanges_OR_NULL-method}
\title{Methods for AutomorphismList-class Objects}
\usage{
as.AutomorphismList(x, grs = GRanges(), ...)

\S4method{as.AutomorphismList}{GRangesList,GRanges_OR_NULL}(x, grs = GRanges(), ...)

\S4method{as.AutomorphismList}{list,GRanges_OR_NULL}(x, grs = GRanges(), ...)
}
\arguments{
\item{x}{A \code{\link[S4Vectors]{DataFrame}} or a
\code{\link{automorphisms}} class object.}

\item{grs}{A \code{\link[GenomicRanges]{GRanges-class}} object.}

\item{...}{Not in use yet.}
}
\value{
The returned an AutomorphismList-class object.
}
\description{
Several methods are available to be applied on
\code{\link{Automorphism-class}} and \code{\link{AutomorphismList-class}}
objects.
}
\seealso{
\code{\link{automorphism_bycoef}}, \code{\link{automorphisms}}
}
