% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{as.AutomorphismList}
\alias{as.AutomorphismList}
\alias{as.AutomorphismList,GRangesList,GRanges_OR_NULL-method}
\alias{as.AutomorphismList,list,GRanges_OR_NULL-method}
\alias{names,AutomorphismList-method}
\alias{names<-,AutomorphismList-method}
\alias{as.list,AutomorphismList-method}
\alias{show,AutomorphismList-method}
\alias{show-AutomorphismList}
\title{Methods for Automorphism-class and AutomorphismList-class Objects}
\usage{
as.AutomorphismList(x, grs = GRanges(), ...)

\S4method{as.AutomorphismList}{GRangesList,GRanges_OR_NULL}(x, grs = GRanges(), ...)

\S4method{as.AutomorphismList}{list,GRanges_OR_NULL}(x, grs = GRanges(), ...)

\S4method{names}{AutomorphismList}(x)

\S4method{names}{AutomorphismList}(x) <- value

\S4method{as.list}{AutomorphismList}(x)

\S4method{show}{AutomorphismList}(object)
}
\arguments{
\item{x}{A \code{\link[S4Vectors]{DataFrame}} or a
\code{\link{automorphisms}} class object.}

\item{object}{An object from \code{\link{AutomorphismList-class}}.}

\item{gr}{A \code{\link[GenomicRanges]{GRanges-class}} object.}
}
\value{
The returned value depends on (and it is understood) from the
method called.
}
\description{
Several methods are available to be applied on
\code{\link{Automorphism-class}} and \code{\link{AutomorphismList-class}}
objects.

Replace the element names in AutomorphismList-class object
}
\section{Automorphism-class methods}{

\subsection{as(from, "Automorphism"):}{

Permits the transformation of a \code{\link[base]{data.frame}} or a
\code{\link[S4Vectors]{DataFrame-class}} object into
\emph{\strong{Automorphism-class}} object if the proper columns are
provided.

Methods from \code{\link[GenomicRanges]{GRanges-class}} can also be
applied
}
}

\section{AutomorphismList-class methods}{

\subsection{as.AutomorphismList(x):}{

\emph{\strong{as.AutomorphismList}} function transform a list of
\code{\link[GenomicRanges]{GRanges-class}}, a
\code{\link[GenomicRanges]{GRangesList-class}}, a list of
\code{\link[base]{data.frame}} or a
\code{\link[S4Vectors]{DataFrame-class}}
objects into a \emph{\strong{AutomorphismList-class}} object.
}

\subsection{unlist(x)}{

It transforms a AutomorphismList-class object into an Automorphism-class
object.
}

\subsection{as.list(x)}{

It transforms a list of Automorphism-class objects into an
AutomorphismList-class object.
}

\subsection{as(x, "GRangesList")}{

It transforms a 'GRangesList' of Automorphism-class objects into an
AutomorphismList-class object.
}

\subsection{names(x)}{

To get the element's names from an AutomorphismList-class object.
}

\subsection{names(x) <- value}{

To assign names to the element from an AutomorphismList-class
object.
}
}

\section{AutomorphismByCoefList-class methods}{

\subsection{unlist(x):}{

It transforms a AutomorphismByCoefList-class object into an
AutomorphismByCoef-class object.
}

\subsection{as(x, "AutomorphismByCoefList")}{

It transforms a 'list' of AutomorphismByCoef-class object into an
AutomorphismByCoefList-class object.
}
}

\examples{
## Load datasets
data(autm, brca1_autm)

## Transforming a list of Automorphisms into an AutomorphismList object
lista <- list(human = brca1_autm[[1]], gorilla = brca1_autm[[2]])
as.AutomorphismList(lista)

## Alternatively we can set
aut <- as.list(brca1_autm[1:2])
class(aut)

## And reverse it
aut <- as.AutomorphismList(aut)
aut

## Let's get the element names from object 'aut'
names(aut)

## Let's assign new names
names(aut) <- c("human_1", "gorilla_1")
names(aut)

## Transforming a GRangesList of Automorphisms into an AutomorphismList
## object
lista <- as(lista, "GRangesList")
as.AutomorphismList(lista)

## Transform a AutomorphismList-class object into an Automorphism-class
## object 
unlist(brca1_autm[1:2])

## Let's transform a AutomorphismByCoefList-class object into an 
## AutomorphismByCoef-class object
data("autby_coef")
unlist(autby_coef[1:2])

## Herein a 'list' object of AutomorphismByCoef-class objects
lista <- list(human = autby_coef[[1]], gorilla = autby_coef[[2]])

## Let's transform the the last list 'lista' into an
## AutomorphismByCoefList-class object
aut <- as(lista, "AutomorphismByCoefList")
aut

## Let's get the element names from object 'aut'
names(aut)

## Let's assign new names
names(aut) <- c("human_1", "gorilla_1")
names(aut)
## Load a DNA sequence alignment
data("brca1_autm", package = "GenomAutomorphism")

## The list of the first three elements
autm_list <- as.list(brca1_autm[1:3])
autm_list
}
\seealso{
\code{\link{automorphism_bycoef}}, \code{\link{automorphisms}}
}
\keyword{internal}
