% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Automorphism-class}
\alias{Automorphism-class}
\alias{Automorphism}
\alias{DataFrame_OR_data.frame-class}
\alias{AutomorphismList-class}
\alias{AutomorphismList}
\alias{valid.AutomorphismList}
\alias{AutomorphismByCoefList-class}
\alias{AutomorphismByCoefList}
\title{A class definition to store codon automorphisms in given as an
Abelian group representation.}
\usage{
valid.AutomorphismList(x)
}
\arguments{
\item{x}{A 'AutomorphismList object'}
}
\value{
Given the slot values, it defines an Automorphism-class object.

AutomorphismByCoefList-class definition.
}
\description{
Two classes are involved in to storing codon automorphisms:
\emph{\strong{Automorphism-class}} and
\emph{\strong{AutomorphismList-class}}.

A class definition to store list of Automorphism class objects
derived from the pairwise automorphism estimation from pairwise
alignments.
}
\details{
An \emph{\strong{Automorphism-class}} object has six
columns: "seq1", "seq2","coord1", "coord2", "autm", and "cube". See the
examples for function \code{\link{automorphisms}}. Observe that as the
\emph{\strong{Automorphism-class}} inherits from
\code{\link[GenomicRanges]{GRanges-class}} the transformation starting
from a \code{\link[GenomicRanges]{GRanges-class}} object into an
\emph{\strong{Automorphism-class}} is straightforward.

However, the transformation starting from a \code{\link[base]{data.frame}}
or a \code{\link[S4Vectors]{DataFrame-class}} object \eqn{"x"} requires for
the creation of an additional \code{\link[GenomicRanges]{GRanges-class}}
object, which by default will have the argument seqnames = "1", strand =
"+", start/end = 1:nrow(x), length = nrow(x). These details must be keep in
mind to prevent fundamental errors in the downstream analyses.

\strong{AutomorphismByCoefList-class} has the following methods:
\subsection{as('from', "AutomorphismByCoefList")}{

Where 'from' is a list of \strong{AutomorphismByCoef-class}.
}

\subsection{unlist(x)}{

Where 'x' is a an \strong{AutomorphismByCoefList-class} object.
}
}
\keyword{internal}
