% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{AutomorphismByCoef-class}
\alias{AutomorphismByCoef-class}
\alias{AutomorphismByCoef}
\alias{valid.AutomorphismByCoef}
\alias{valid.AutomorphismByCoefList}
\title{A class definition to store conserved gene/genomic regions found
in a MSA.}
\usage{
valid.AutomorphismByCoef(x)

valid.AutomorphismByCoefList(x)
}
\arguments{
\item{x}{A 'AutomorphismByCoefList object'}
}
\value{
AutomorphismByCoef-class definition.
}
\description{
A class definition to store conserved gene/genomic regions found
in a MSA.

Valid AutomorphismByCoef mcols

Valid AutomorphismByCoefList mcols
}
\keyword{internal}
