% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-methods.R
\name{seq2granges}
\alias{seq2granges}
\alias{seq2granges,DNAStringSet_OR_NULL-method}
\alias{base_seq2string_set}
\alias{base_seq2string_set,BaseSeq-method}
\title{DNA sequences to GRanges of bases and the reverse.}
\usage{
seq2granges(base = NULL, filepath = NULL, ...)

\S4method{seq2granges}{DNAStringSet_OR_NULL}(
  base = NULL,
  filepath = NULL,
  start = NA,
  end = NA,
  chr = 1L,
  strand = "+",
  seq_alias = NULL
)

base_seq2string_set(x, ...)

\S4method{base_seq2string_set}{BaseSeq}(x)
}
\arguments{
\item{filepath}{A character vector containing the path to a file in
\emph{\strong{fasta}} format to be read. This argument must be given if
\emph{codon & base} arguments are not provided.}

\item{...}{Not in use.}

\item{start, end, chr, strand}{Optional parameters required to build a
\code{\link[GenomicRanges]{GRanges-class}}. If not provided the default
values given for the function definition will be used.}

\item{x}{A 'BaseSeq' class object.}
}
\value{
\subsection{seq2granges}{

This function returns a \link{BaseGroup} object carrying the DNA sequence(s), one
base per ranges. A \link{BaseGroup} class object inherits from
\code{\link[GenomicRanges]{GRanges-class}}.
}

\subsection{base_seq2string_set}{

This function returns a \code{\link[Biostrings]{DNAStringSet-class}}.
}
}
\description{
Functions 'seq2granges' and 'base_seq2string_set' are addressed to transform
DNA sequences into \code{\link[GenomicRanges]{GRanges-class}} and the
reverse transformation, respectively.
}
\details{
For the sake of brevety the metacolumns from the object returned by
function 'seq2granges' are named as 'S1', 'S2', 'S3', and so on. The
original DNA sequence alias are stored in the slot named 'seq_alias'.
(see examples).
}
\examples{
## Load a multiple sequence alignment (MSA) of primate BRCA1 DNA repair 
## genes 
data(brca1_aln2, package = "GenomAutomorphism")
brca1_aln2

## Get BaseSeq-class object
gr <- seq2granges(brca1_aln2)
gr

## Transform the BaseSeq-class object into a DNAStringSet-class object
str_set <- base_seq2string_set(gr)
str_set

## Recovering the original MSA
require(Biostrings)
DNAMultipleAlignment(as.character(str_set))

}
\seealso{
\href{https://github.com/genomaths/GenomeAlgebra_SymmetricGroup}{Symmetric Group of the Genetic-Code Cubes.}

\link{base_coord} and \link{codon_coord}.
}
\author{
Robersy Sanchez \url{https://genomaths.com}
}
