% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_brca1_autm.R
\docType{data}
\name{brca1_autm}
\alias{brca1_autm}
\title{Automorphisms between DNA Sequences from Primate BRCA1 Genes}
\format{
\code{\link{AutomorphismList}} class object.
}
\usage{
brca1_autm
}
\description{
This is a \code{\link{AutomorphismList}} object carrying a list of pairwise
automorphisms between the DNA sequences from the MSA of primate BRCA1
DNA repair gene. The automorphisms were estimated from the
\code{\link{brca1_aln}} MSA with function \code{\link{autZ64}}.
}
\examples{
data(brca1_autm, package = "GenomAutomorphism")
brca1_autm

}
\seealso{
\link{brca1_autm2}, \link{brca1_aln}, \link{brca1_aln2}, and \link{covid_autm}.
}
\author{
Robersy Sanchez \url{https://genomaths.com}
}
\keyword{datasets}
