% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{CodonGroup-class}
\alias{CodonGroup-class}
\alias{CodonGroup}
\alias{valid.CodonGroup.mcols}
\title{A class definition to store codon automorphisms in given in the
Abelian group representation.}
\usage{
valid.CodonGroup.mcols(x)
}
\arguments{
\item{x}{A 'CodonGroup' object}
}
\value{
Given the slot values define a CodonGroup-class.
}
\description{
A class definition to store codon automorphisms in given in the
Abelian group representation.

Valid CodonGroup mcols
}
\seealso{
\code{\link{automorphisms}}
}
\keyword{internal}
