% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{CodonMatrix}
\alias{CodonMatrix}
\title{Constructor of CodonMatrix-class}
\usage{
CodonMatrix(object, group, cube, seq_alias = NULL)
}
\arguments{
\item{object}{A \code{\link[GenomicRanges]{GRanges-class}} object.}

\item{group}{The name of the base group, 'Z4' or 'Z5'.}

\item{cube}{The name of the genetic-code cube applied to get the
codon coordinates.}

\item{seq_alias}{The 'alias' given to the codon sequence.}
}
\description{
Constructor of CodonMatrix-class
}
\seealso{
\link{base_coord} and \link{codon_coord}.
}
