% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_phychem.R
\name{dna_phychem}
\alias{dna_phychem}
\alias{dna_phychem,character-method}
\alias{dna_phychem,DNAStringSet_OR_DNAMultipleAlignment-method}
\title{DNA numerical matrix}
\usage{
dna_phychem(seqs, ...)

\S4method{dna_phychem}{character}(
  seqs,
  phychem = list(A = NULL, T = NULL, C = NULL, G = NULL, N = NULL)
)

\S4method{dna_phychem}{DNAStringSet_OR_DNAMultipleAlignment}(
  seqs,
  phychem = list(A = NULL, T = NULL, C = NULL, G = NULL, N = NULL),
  index_name = NULL,
  ...
)
}
\arguments{
\item{...}{Not in use.}

\item{index_name}{Optional. Name of breve description of the base
physicochemical property applied to represent the DNA sequence.}

\item{base}{A character vector, string , or a dataframe of letters from the
DNA/RNA alphabet.}

\item{num.cores, tasks}{Parameters for parallel computation using package
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to
use, i.e. at most how many child processes will be run simultaneously
(see \code{\link[BiocParallel]{bplapply}} and the number of tasks per job
(only for Linux OS).}
}
\description{
This function applies the numerical indices representing various
physicochemical and biochemical properties of DNA bases. As results, DNA
sequences are represented as numerical vectors which can be subject of
further downstream statistical analysis and digital signal processing.
}
\examples{
## Let's create DNAStringSet-class object 
base <- DNAStringSet(x = c( seq1 ='ACGTGATCAAGT', 
                            seq2 = 'GTGTGATCCAGT', 
                            seq3 = 'TCCTGATCAGGT'))


dna_phychem(seqs = base,
            phychem <- list('A' = 0.87, 'C' = 0.88, 'T' = 0.82,
                            'G' = 0.89, 'N' = NA),
            index_name = "Proton-Affinity")

}
\seealso{
\link{peptide_phychem_index}
}
\author{
Robersy Sanchez \url{https://genomaths.com}
}
