% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_dist.R
\docType{data}
\name{binary_seq}
\alias{binary_seq}
\alias{hamming_dist}
\title{Hamming Distance}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 250 columns.
}
\usage{
binary_seq

hamming_dist(
  m1,
  m2 = NULL,
  return_matrix = FALSE,
  num.cores = 1L,
  tasks = 0L,
  verbose = TRUE
)
}
\arguments{
\item{m1, m2}{A \code{\link[IRanges]{AtomicList-class}} (NumericList), a
\code{\link{encNumSignalList-class}} or a data.frame where each row
contains a vector of 0s and 1s, and the row-names are the sequence names.
By default m2 is NULL.}

\item{return_matrix}{logical(1). Whether to return a lower distance matrix
when m2 = NULL. If TRUE, the diagonal and the upper matrix are fielded with
zeros and with lower matrix with the distance values. If FALSE, then a vector
with the distance values for the matrix cells with coordinates ((j + 1):l, j)
is returned (j = 1, ... l - 1, and l = nrow(m1)).}

\item{num.cores, tasks}{Parameters for parallel computation using package
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to
use, i.e. at most how many child processes will be run simultaneously
(see \code{\link[BiocParallel]{bplapply}} and the number of tasks per job
(only for Linux OS).}

\item{verbose}{if TRUE, prints the function log to stdout.}
}
\value{
Depending on the provided data, it will return a lower distance
matrix, a vector with the distance values for the matrix cells with
coordinates ((j + 1):l, j) (j = 1, ... l - 1, and l = nrow(m1)) or a pairwise
distance vector between each vector k from m1 and the corresponding vector k
from m2.
}
\description{
This function compute de Hamming distance between binary
vectors.

\subsection{Compute the distance matrix and return the lower diagonal matrix}{
}

\subsection{as a vector}{

hamming_dist(m1 = binary_seq, return_matrix = FALSE)
}
}
\details{
A pairwise distance between vectors from m1 and m2 is computed when
both arguments, m1 and m2,  are provided. Otherwise, when m2 = NULL, the
distance matrix for vectors from m1 is computed.
}
\examples{
### Generating a matrix carrying three binary sequence of 0s and 1s.
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
\keyword{datasets}
