% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{CodonMatrix-class}
\alias{CodonMatrix-class}
\alias{ListCodonMatrix}
\alias{ListCodonMatrix-class}
\alias{valid.ListCodonMatrix}
\title{A Convenient Class to Store a Codon Coordinate in given
Genetic Code cube.}
\usage{
ListCodonMatrix(object, cube, group, seq_alias = NULL)

valid.ListCodonMatrix(x)
}
\arguments{
\item{object}{A list of CodonMatrix-class objects}

\item{x}{A 'ListCodonMatrix-class' object}
}
\description{
A Convenient Class to Store a Codon Coordinate in given
Genetic Code cube.

A Convenient Class to Store Codon Coordinates in given
Genetic Code cube.

Constructor of ListCodonMatrix-class

Valid ListCodonMatrix-class object
}
\keyword{internal}
