% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{MatrixSeq-class}
\alias{MatrixSeq-class}
\alias{MatrixSeq}
\alias{show,MatrixSeq-method}
\alias{show-MatrixSeq}
\title{Definition of MatrixSeq-class}
\usage{
\S4method{show}{MatrixSeq}(object)
}
\arguments{
\item{object}{An object from 'MatrixSeq' class}
}
\value{
Given the slot values, it defines a MatrixList-class.

Print/show of a MatrixSeq-class object.
}
\description{
This is a very simple flexible class to store DNA and
aminoacid aligned sequences together with their physicochemical properties.
That is, a place where each aminoacid or codon from the sequence is
represented by numerical value from a physicochemical index.
}
\details{
\itemize{
\item{\strong{seqs}: }{A string character vector of DNA or aminoacid
sequences.}
\item{\strong{matrix}: }{A numerical matrix carrying the specified
aminoacid physicochemical indices for aminoacid in the DNA or aminoacid
sequences.}
\item{\strong{names}: }{Alias/names/IDs DNA or aminoacid sequences.}
\item{\strong{aaindex}: }{Aminoacid index database where the
physicochemical index can be found.}
\item{\strong{phychem}: }{Description of the physicochemical index applied
to represent the DNA or aminoacid sequences.}
\item{\strong{accession}: }{Accession number or ID of the applied
physicochemical index in the database.}
}
}
\keyword{internal}
