% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlineq.R
\name{modlineq}
\alias{modlineq}
\title{Modular System of Linear Equation Solver (MLE)}
\usage{
modlineq(a, b, n)
}
\arguments{
\item{a}{An integer or a vector of integers.}

\item{b}{An integer or a vector of integers.}

\item{n}{An integer or a vector of integers.}
}
\value{
A numerical vector. If for some coordinate the equation has not
solution in their definition domain it will return 0 for such coordinate.
}
\description{
If \eqn{a, b}, and  \eqn{c} are integer vectors, this function
try to find, at each coordinate, the solution of the MLE
\eqn{a x = b}  mod \eqn{n}. If the MLE \eqn{a x = b mod n} has not
solutions (see \code{\link[numbers]{modlin}}), the value reported for the
coordinate will be 0 and the corresponding translation.
}
\details{
For \eqn{a, b}, and \eqn{c} integer scalars, it is just a
wrapper function to call \code{\link[numbers]{modlin}}.
}
\examples{
## The MLE 8x = 54 mod 64 is not solvable; hence the first vector 
## coordinate returns 0.
modlineq(a = c(8,9), b = c(54, 34), n = c(64,64))
}
