% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{valid.AutomorphismByCoef}
\alias{valid.AutomorphismByCoef}
\alias{AutomorphismByCoefList-class}
\alias{AutomorphismByCoefList}
\alias{valid.AutomorphismByCoefList}
\title{Valid AutomorphismByCoef mcols}
\usage{
valid.AutomorphismByCoef(x)

valid.AutomorphismByCoefList(x)
}
\arguments{
\item{x}{A 'AutomorphismByCoefList object'}
}
\value{
AutomorphismByCoefList-class definition.
}
\description{
Valid AutomorphismByCoef mcols

A class definition for a list of AutomorphismByCoef class objects.

Valid AutomorphismByCoefList mcols
}
\details{
\strong{AutomorphismByCoefList-class} has the following methods:
\subsection{as('from', "AutomorphismByCoefList")}{

Where 'from' is a list of \strong{AutomorphismByCoef-class}.
}

\subsection{unlist(x)}{

Where 'x' is a an \strong{AutomorphismByCoefList-class} object.
}
}
\keyword{internal}
