% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlineq.R
\name{modlineq}
\alias{modlineq}
\title{Modular System of Linear Equation Solver (MLE)}
\usage{
modlineq(a, b, n, no.sol = 0L)
}
\arguments{
\item{a}{An integer or a vector of integers.}

\item{b}{An integer or a vector of integers.}

\item{n}{An integer or a vector of integers.}

\item{no.sol}{Values to return when the equation is not solvable or yield
the value 0. Default is 0.}
}
\value{
If the solution is exact, then a numerical vector will be returned,
otherwise, if there is not exact solution for some coordinate, the a list
carrying the element on the diagonal matrix and the tranlation vector will
be returned.
}
\description{
If \eqn{a, b}, and  \eqn{c} are integer vectors, this function
try to find, at each coordinate, the solution of the MLE
\eqn{a x = b}  mod \eqn{n}. If the MLE \eqn{a x = b mod n} has not
solutions (see \code{\link[numbers]{modlin}}), the value reported for the
coordinate will be 0 and the corresponding translation.
}
\details{
For \eqn{a, b}, and \eqn{c} integer scalars, it is just a
wrapper function to call \code{\link[numbers]{modlin}}.
}
\examples{
## The MLE 8x = 54 mod 64 is not solvable; hence the first vector 
## coordinate returns 0.
modlineq(a = c(8,9), b = c(54, 34), n = c(64,64))
}
