% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_phychem_index.R
\name{peptide_phychem_index}
\alias{peptide_phychem_index}
\alias{peptide_phychem_index,character-method}
\alias{peptide_phychem_index,DNAStringSet-method}
\title{Amino acid mutation matrix}
\usage{
peptide_phychem_index(aa, ...)

\S4method{peptide_phychem_index}{character}(
  aa,
  acc = NULL,
  aaindex = NA,
  userindex = NULL,
  alphabet = c("AA", "DNA"),
  ...
)

\S4method{peptide_phychem_index}{DNAStringSet}(
  aa,
  acc = NULL,
  aaindex = NA,
  userindex = NULL,
  num.cores = 1L,
  tasks = 0L,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{...}{Not in use.}

\item{acc}{Accession id for a specified mutation or contact potential
matrix.}

\item{aaindex}{Database where the requested accession id is locate and from
where the aminoacid indices can be obtained. The possible values are:
"aaindex2" or "aaindex3".}

\item{userindex}{User provided aminoacid indices. This can be a numerical
vector or a matrix (20 x 20). If a numerical matrix is provided, then the
aminoacid indices are computes as column averages.}

\item{alphabet}{Whether the alphabet is from the 20 aminoacid (AA) or
four (DNA)/RNA base alphabet. This would prevent mistakes, i.e.,
the strings "ACG" would be a base-triplet on the DNA alphabet or simply
the amino acid sequence of alanine, cysteine, and glutamic acid.}

\item{num.cores, tasks}{Parameters for parallel computation using package
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to
use, i.e. at most how many child processes will be run simultaneously
(see \code{\link[BiocParallel]{bplapply}} and the number of tasks per job
(only for Linux OS).}

\item{verbose}{If TRUE, prints the function log to stdout.}

\item{acc_list}{Logical. If TRUE, then the list of available matrices ids
and index names is returned.}
}
\value{
Depending on the user specifications, a mutation or contact
potential matrix, a list of available matrices (indices) ids or index
names can be returned. More specifically:

\itemize{
\item{\strong{aa_mutmat}: }{Returns an aminoacid mutation matrix or
a statistical protein contact potentials matrix.}
\item{\strong{aa_index}: }{Returns the specified aminoacid physicochemical
indices.}
}
}
\description{
This function applies the numerical indices representing various
physicochemical and biochemical properties of amino acids and
pairs of amino acids to whole DNA protein-coding or to aminoacid sequences.
As results, DNA protein-coding or the aminoacid sequences are represented
as numerical vectors which can be subject of further
downstream statistical analysis and digital signal processing.
}
